changed http:// to // so they will work on both http and https pages
Ran into this problem on cloudfront hosted site. That page was blank
because the resources were unsecured and we require https with
cloudfront. However, in development the S3 bucket is not secured and
https URLs would also cause a mixed-content error. Removing the protocol
moniker makes the resource use whatever protocl the page was loaded with.